simplify type export
