+++ lib/index.d.ts
@@ -1,6 +1,6 @@
 import Onyx, {OnyxUpdate} from './Onyx';
-import {CustomTypeOptions, OnyxCollection, OnyxEntry, ConnectOptions} from './types';
+import {CustomTypeOptions, OnyxCollection, OnyxEntry} from './types';
 import withOnyx from './withOnyx';
 
 export default Onyx;
-export {CustomTypeOptions, OnyxCollection, OnyxEntry, OnyxUpdate, withOnyx, ConnectOptions};
+export {CustomTypeOptions, OnyxCollection, OnyxEntry, OnyxUpdate, withOnyx};
