+++ lib/types.d.ts
@@ -1,5 +1,4 @@
 import {IsEqual, Merge} from 'type-fest';
-import {Component} from 'react'
 
 /**
  * Represents a deeply nested record. It maps keys to values,
@@ -183,41 +182,6 @@ type OnyxEntry<TOnyxValue> = TOnyxValue | null;
  */
 type OnyxCollection<TOnyxValue> = OnyxEntry<Record<string, TOnyxValue | null>>;
 
-/**
- * Represents the base options used in `Onyx.connect()` method.
- */
-type BaseConnectOptions = {
-    statePropertyName?: string;
-    withOnyxInstance?: Component;
-    initWithStoredValues?: boolean;
-};
-
-/**
- * Represents the options used in `Onyx.connect()` method.
- * The type is built from `BaseConnectOptions` and extended to handle key/callback related options.
- * It includes two different forms, depending on whether we are waiting for a collection callback or not.
- *
- * If `waitForCollectionCallback` is `true`, it expects `key` to be a Onyx collection key and `callback` will be triggered with the whole collection
- * and will pass `value` as an `OnyxCollection`.
- *
- *
- * If `waitForCollectionCallback` is `false` or not specified, the `key` can be any Onyx key and `callback` will be triggered with updates of each collection item
- * and will pass `value` as an `OnyxEntry`.
- */
-type ConnectOptions<TKey extends OnyxKey> = BaseConnectOptions &
-    (
-        | {
-              key: TKey extends CollectionKey ? TKey : never;
-              callback?: (value: OnyxCollection<KeyValueMapping[TKey]>) => void;
-              waitForCollectionCallback: true;
-          }
-        | {
-              key: TKey;
-              callback?: (value: OnyxEntry<KeyValueMapping[TKey]>, key: TKey) => void;
-              waitForCollectionCallback?: false;
-          }
-    );
-
 export {
     CollectionKey,
     CollectionKeyBase,
@@ -229,6 +193,4 @@ export {
     OnyxEntry,
     OnyxKey,
     Selector,
-    BaseConnectOptions,
-    ConnectOptions
 };
